+++ api/context/account/domain/model/user.go
@@ -1,4 +1,4 @@
-package user
+package model
 
 import (
 	"lmm/api/domain/model"
