+++ api/context/account/usecase/signup_test.go
@@ -1,7 +1,7 @@
-package user
+package usecase
 
 import (
-	repo "lmm/api/domain/repository/user"
+	"lmm/api/context/account/domain/repository"
 	"lmm/api/testing"
 )
 
@@ -12,7 +12,7 @@ func TestSignUp(t *testing.T) {
 	auth := Auth{Name: "foobar", Password: "1234"}
 	requestBody := testing.StructToRequestBody(auth)
 
-	id, err := New(repo.New()).SignUp(requestBody)
+	id, err := New(repository.New()).SignUp(requestBody)
 	tester.NoError(err)
 	tester.Is(uint64(1), id)
 }
@@ -22,8 +22,9 @@ func TestSignUp_Duplicate(t *testing.T) {
 	tester := testing.NewTester(t)
 
 	auth := Auth{Name: "foobar", Password: "1234"}
-	New(repo.New()).SignUp(testing.StructToRequestBody(auth))
-	id, err := New(repo.New()).SignUp(testing.StructToRequestBody(auth))
+	repo := repository.New()
+	New(repo).SignUp(testing.StructToRequestBody(auth))
+	id, err := New(repo).SignUp(testing.StructToRequestBody(auth))
 	tester.Error(err)
 	tester.Is(ErrDuplicateUserName, err)
 	tester.Is(uint64(0), id)
